Добавлен абстрактный класс PixelFilter
