+++ PhotoEnhancer/PhotoEnhancer/Filters/GrayScaleFilter.cs
@@ -7,29 +7,15 @@ using System.Windows.Forms;
 
 namespace PhotoEnhancer
 {
-    public class GrayScaleFilter : IFilter
+    public class GrayScaleFilter : PixelFilter
     {
-        public ParameterInfo[] GetParametersInfo() => new ParameterInfo[0];
+        public override ParameterInfo[] GetParametersInfo() => new ParameterInfo[0];
 
-        public Photo Process(Photo original, double[] parameters)
+        public override Pixel ProcessPixel(Pixel pixel, double[] parameters)
         {
-            var newPhoto = new Photo(original.Width, original.Height);
+            var lightness = pixel.R * 0.3 + pixel.G * 0.6 + pixel.B * 0.1;
 
-            for (var x = 0; x < original.Width; x++)
-                for (var y = 0; y < original.Height; y++)
-                {
-                    var lightness =
-                        original[x, y].R * 0.3 +
-                        original[x, y].G * 0.6 +
-                        original[x, y].B * 0.1;
-
-                    Pixel p = new Pixel(lightness, lightness, lightness);
-
-                    newPhoto[x, y] = p;
-                }
-                    
-
-            return newPhoto;          
+            return new Pixel(lightness, lightness, lightness);
         }
 
         public override string ToString() => "Оттенки серого";
